+++ Client.Legacy/FluxConnectionOptions.cs
@@ -3,7 +3,13 @@ using System;
 namespace InfluxDB.Client.Flux
 {
     public class FluxConnectionOptions
-    {
+    {   
+        public enum AuthenticationType
+        {
+            UrlQueryParameters,
+            BasicAuthentication
+        }
+
         public string Url { get; private set; }
 
         public TimeSpan Timeout { get; private set; }
@@ -11,21 +17,26 @@ namespace InfluxDB.Client.Flux
         public string Username { get; }
         public char[] Password { get; }
 
+        public AuthenticationType Authentication { get; }
+
         public FluxConnectionOptions(string url) : this(url, TimeSpan.FromSeconds(60))
         {
         }
 
-        public FluxConnectionOptions(string url, string username = "", char[] password = null) : this(url,
-            TimeSpan.FromSeconds(60), username, password)
+        public FluxConnectionOptions(string url, string username = "", char[] password = null, 
+                        AuthenticationType authentication = AuthenticationType.UrlQueryParameters) : this(url,
+            TimeSpan.FromSeconds(60), username, password, authentication)
         {
         }
 
-        public FluxConnectionOptions(string url, TimeSpan timeout, string username = "", char[] password = null)
+        public FluxConnectionOptions(string url, TimeSpan timeout, string username = "", char[] password = null, 
+                        AuthenticationType authentication = AuthenticationType.UrlQueryParameters)
         {
             Url = url;
             Timeout = timeout;
             Username = username;
             Password = password;
+            Authentication = authentication;
         }
     }
 }